Remove callback in initialize and migrate core startup to package
