+++ pkg/startup/cloudprovider.go
@@ -1,0 +1,30 @@
+/*
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package startup
+
+import (
+	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
+	"sigs.k8s.io/controller-runtime/pkg/manager"
+
+	"github.com/aws/karpenter/pkg/cloudprovider"
+	cloudprovidermetrics "github.com/aws/karpenter/pkg/cloudprovider/metrics"
+)
+
+func Decorate(cloudProvider cloudprovider.CloudProvider, m manager.Manager) cloudprovider.CloudProvider {
+	if hp, ok := cloudProvider.(HealthCheck); ok {
+		utilruntime.Must(m.AddHealthzCheck("cloud-provider", hp.LivenessProbe))
+	}
+	return cloudprovidermetrics.Decorate(cloudProvider)
+}
